all app content implemented in separated component
