+++ src/components/App/App.tsx
@@ -1,21 +1,10 @@
 import React from 'react';
 
-import { Box, Container, styled } from '@mui/material';
-import { TodoAppTitle } from '../Title/Title';
-import { ThemeSwitchMenu } from '../ThemeSwitchMenu/ThemeSwitchMenu';
-
-const StyledBox = styled(Box)({
-  width: '100%',
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-});
+import { Container } from '@mui/material';
+import { AppContent } from '../AppContent';
 
 export const App: React.FC = () => (
   <Container>
-    <StyledBox>
-      <TodoAppTitle />
-      <ThemeSwitchMenu />
-    </StyledBox>
+    <AppContent />
   </Container>
 );
